% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scpdsi.R
\name{pdsi}
\alias{pdsi}
\title{Calculate the (sc)PDSI}
\usage{
pdsi(
  indat,
  AWC = NULL,
  start = NULL,
  end = NULL,
  cal_start = NULL,
  cal_end = NULL,
  sc = TRUE
)
}
\arguments{
\item{AWC}{Available soil water capacity of the soil layer [mm]. Default 100 mm.}

\item{start}{Integer. Start year of the PDSI to be calculate default 1.}

\item{end}{Integer. End year of the PDSI to be calculate.}

\item{cal_start}{Integer. Start year of the calibrate period. Default is start year.}

\item{cal_end}{Integer. End year of the calibrate period. Default is end year.}

\item{sc}{Bool. Should use the self-calibrating procedure to calculate the climatical
coefficient (K2 and duration coefficients). If not it would use the default
parameters of Palmer (1965).}

\item{P}{Monthly precipitation series without NA [mm]. Can be a time series.}

\item{PE}{Monthly potential evapotranspiration corresponding to the precipitation
series. Can be calculated by the Penman-Monteith or the Thonthwate
equation [mm].}
}
\value{
This function return a vector of scpdsi.
}
\description{
Calculating the monthly conventional Palmer Drought Severity Index
             (PDSI) and Self-calibrating PDSI (scPDSI) using the precipitation
             and potential evapotranspiration.
}
\details{
The Palmer Drought Severity Index (PDSI), proposed by Palmer (1965), is a
widely used drought indicator to quantify the long-term drought conditions, for
an area at a certain time. The PDSI is a semi-physical based drought index calculated
using the precipitation and potential evapotranspiration data, based on a simple
two-layer bucket water balance model. Conventionally, the constants to calculate
the PDSI were firstly empirically derived by using the meteorological records in
Kansas and Iowa in middle US with a semi-arid climate conditions, therefore the
conventional PDSI usually could not satisfactorily represent the drought conditions
for other areas around the world, which also makes spatial comparisons of PDSI values
difficult.

For this, Wells et al. (2004) proposed a self-calibrating Palmer Drought Severity
Index (scPDSI). The scPDSI could automatically adjust the empirical constants
in the PDSI computation with dynamically calculated values. Several works have
proved that the scPDSI performs better in spatially comparison than the conventional
PDSI. For more details please see the works of Wells et al. (2004).

This function could also calculate the conventional PDSI with revised constants.
Before the scPDSI appeared, the PDSI calculation has also been revised focusing
on the local climate characteristics in many area around the world. Those
constants could be reset by using the global options before calculating the PDSI
e.g.:

\preformatted{
options(PDSI.coe.K1.1 = 1.6)
}

And calculation in the PDSI would be:
\deqn{Ki0=coeK11 * lg(((PEi+Ri+ROi)/(Pi+Li)+coeK12)/Di)+coeK13}
\deqn{Ki=coeK2/(\sum Dj*Kj) * Ki}
\deqn{X[i]=p*X[i-1]+q*Z[i]}

\eqn{coeK11}, \eqn{coeK12}, \eqn{coeK13}, \eqn{coeK2}, \eqn{p}, and \eqn{q} are
corresponding to \code{PDSI.coe.K1.1}, \code{PDSI.coe.K1.2}, \code{PDSI.coe.K1.3},
\code{PDSI.coe.K2}, \code{PDSI.p}, and \code{PDSI.q}, respectively.

For example, in a national standard in China about meteorological drought level
(GB/T 20481-2017), the PDSI was revised by re-adjust the constants in the
calculation. To use the PDSI based on this standard should set the global
options of R as follows:

\preformatted{
options(PDSI.coe.K1.1 = 1.6)
options(PDSI.coe.K1.3 = 0.4)
options(PDSI.coe.K2 = 16.84)
options(PDSI.p = 0.755)
options(PDSI.q = 1/1.63)
}
}
\examples{
library(scPDSIr)


}
\references{
Palmer W., 1965. Meteorological drought. U.s.department of Commerce
            Weather Bureau Research Paper.

            Wells, N., Goddard, S., Hayes, M. J., 2004. A Self-Calibrating Palmer
            Drought Severity Index. Journal of Climate, 17(12):2335-2351.
}
