% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scpdsi.R
\name{pdsi}
\alias{pdsi}
\title{Calculate the (sc)PDSI}
\usage{
pdsi(
  indat,
  AWC = NULL,
  start = NULL,
  end = NULL,
  cal_start = NULL,
  cal_end = NULL,
  sc = TRUE
)
}
\value{
This function return a vector of scpdsi.
}
\description{
Calculating the monthly Self-calibrating PDSI (scPDSI) using the precipitation
             and potential evapotranspiration.
}
\details{
The Palmer Drought Severity Index (PDSI), proposed by Palmer (1965), is a
widely used drought indicator to quantify the long-term drought conditions, for
an area at a certain time. The PDSI is a semi-physical based drought index calculated
using the precipitation and potential evapotranspiration data, based on a simple
two-layer bucket water balance model. Conventionally, the constants to calculate
the PDSI were firstly empirically derived by using the meteorological records in
Kansas and Iowa in middle US with a semi-arid climate conditions, therefore the
conventional PDSI usually could not satisfactorily represent the drought conditions
for other areas around the world, which also makes spatial comparisons of PDSI values
difficult.

For this, Wells et al. (2004) proposed a self-calibrating Palmer Drought Severity
Index (scPDSI). The scPDSI could automatically adjust the empirical constants
in the PDSI computation with dynamically calculated values. Several works have
proved that the scPDSI performs better in spatially comparison than the conventional
PDSI. For more details please see the works of Wells et al. (2004).
}
\examples{

library(scPDSIr)
scpdsi(indat,  start, end, cal_start, cal_end)
Where the dataframe "indat" contains the following variables:

P = Monthly precipitation series without NA [mm]. Can be a time series.

PE = Monthly potential evapotranspiration corresponding to the precipitation
           series. Can be calculated by the Penman-Monteith or the Thonthwate
           equation [mm].
lat = Latitude  (degrees)

lon = Longitude (degrees)

The following are provided as separate arguments:

AWC Available soil Water Capacity of the soil layer [mm]. This will be retrieved automatically
from a 0.1 degree European grid and does not need to be supplied. However 
if using a single site this can be specified if required.

start Integer. Start year of the PDSI to be calculated default 1.

end Integer. End year of the PDSI to be calculated.

cal_start Integer. Start year of the calibrate period. Default is start year.

cal_end Integer. End year of the calibrate period. Default is end year.

}
\references{
Palmer W., 1965. Meteorological drought. U.s.department of Commerce
            Weather Bureau Research Paper.

            Wells, N., Goddard, S., Hayes, M. J., 2004. A Self-Calibrating Palmer
            Drought Severity Index. Journal of Climate, 17(12):2335-2351.
}
